% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlm_fit.R
\name{chlm_fit}
\alias{chlm_fit}
\title{Estimate the parameters of the HLM model.}
\usage{
chlm_fit(y, delta, X, Z, maxit = 1000, epsilon = 1e-08,
  splitE = FALSE)
}
\arguments{
\item{y}{Vector of observations of length \code{n}.}

\item{delta}{Optional logical vector of length \code{n} indicating the censorring status (\code{TRUE}: uncensored, \code{FALSE}: uncensored).}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}

\item{maxit}{Maximum number of iteration of the fitting algorithm (see \strong{Details}).}

\item{epsilon}{Tolerance threshold for termination of the algorithm (see \strong{Details}).}

\item{splitE}{If \code{TRUE}, perform the E-step after each conditional M-step (see \strong{Details}).}
}
\value{
An object of class \code{hlm} with the following elements:
\describe{
  \item{\code{beta}}{The MLE of the mean parameter vector.}
  \item{\code{gamma}}{The MLE of the variance parameter vector.}
  \item{\code{loglik}}{The value of the loglikelihood at the fitted \code{(beta, gamma)}.}
  \item{\code{niter}}{The number of steps taken by the algorithm.}
  \item{\code{error}}{The value of the error at the end of the algorithm.}
}
}
\description{
Estimate the parameters of the HLM model.
}
\details{
The fitting algorithm is an Expectation-Conditional-Maximization (ECM) algorithm extending the alternating weighted-LM/GLM updates of \code{beta} and \code{gamma}, proposed by Smyth??? for the uncensored setting.  The ECM algorithm terminates when either \code{maxit} iterations have been reached, or when
\preformatted{
|ll_curr - ll_prev| / (0.1 + |ll_curr|) < epsilon,
}
where \code{ll_curr} and \code{ll_prev} are the loglikelihood values at the current and previous iterations.

\strong{TODO:}
\itemize{
  \item Input checking.
  \item \code{print}, \code{summary}, \code{vcov} methods.
  \item \code{residual} method.  Perhaps use expected lifetime for the censored observations?
  \item Separate into \code{chlm} and \code{hlm} classes?
}
}
