% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlm_grad.R
\name{chlm_grad}
\alias{chlm_grad}
\title{Gradient of the HLM loglikelihood.}
\usage{
chlm_grad(beta, gamma, y, delta, X, Z)
}
\arguments{
\item{beta}{Mean parameter vector of length \code{p}.}

\item{gamma}{Variance parameter vector of length \code{q}.}

\item{y}{Vector of observations of length \code{n}.}

\item{delta}{Optional logical vector of length \code{n} indicating the censorring status (\code{TRUE}: uncensored, \code{FALSE}: uncensored).}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}
}
\value{
Vector of length \code{p+q} returning the gradient of the loglikelihood at the parameter values.
}
\description{
Gradient of the HLM loglikelihood.
}
