% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlm_grad.R
\name{chlm_hess}
\alias{chlm_hess}
\title{Hessian of the HLM loglikelihood.}
\usage{
chlm_hess(beta, gamma, y, delta, X, Z)
}
\arguments{
\item{beta}{Mean parameter vector of length \code{p}.}

\item{gamma}{Variance parameter vector of length \code{q}.}

\item{y}{Vector of observations of length \code{n}.}

\item{delta}{Optional logical vector of length \code{n} indicating the censorring status (\code{TRUE}: uncensored, \code{FALSE}: uncensored).}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}
}
\value{
Matrix of size \code{(p+q) x (p+q)} containg the Hessian of the loglikelihood at the parameter values.
}
\description{
Hessian of the HLM loglikelihood.
}
