% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlm_fit.R
\name{chlm_loglik}
\alias{chlm_loglik}
\title{Loglikelihood of the HLM model.}
\usage{
chlm_loglik(beta, gamma, y, delta, X, Z)
}
\arguments{
\item{beta}{Mean parameter vector of length \code{p}.}

\item{gamma}{Variance parameter vector of length \code{q}.}

\item{y}{Vector of observations of length \code{n}.}

\item{delta}{Optional logical vector of length \code{n} indicating the censorring status (\code{TRUE}: uncensored, \code{FALSE}: uncensored).}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}
}
\value{
The loglikelihood evaluated at the parameters values (a scalar).
}
\description{
Loglikelihood of the HLM model.
}
