% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm-package.R
\docType{package}
\name{hlm-package}
\alias{hlm}
\alias{hlm-package}
\title{hlm: Inference for a Heteroscedastic Linear Model}
\description{
Provides efficient algorithms to compute the MLE of a heteroscedastic regression model, with linear mean and log-linear variance covariates.  The package accommodates right-censored responses, which are expected in the context of survival analysis for the corresponding heteroscedastic accelerated failure-time model of Wang et al (2019).
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

Authors:
\itemize{
  \item Tian You
  \item Yifan Wang
}

}
\keyword{internal}
