% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm.R
\name{hlm}
\alias{hlm}
\title{Fit the heteroscedastic linear model.}
\usage{
hlm(formula, data, subset, weights, na.action, control, model = TRUE,
  qr = TRUE, x = FALSE, y = FALSE, offset)
}
\arguments{
\item{formula}{An object of class \code{formula} (or one that can be coerced to that class).  See \strong{Details}.}

\item{data}{An optional data frame, list or environment containing the variables in the model.  If not found in \code{data}, the variables are taken from \code{environment(formula)}.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{Currently ignored.}

\item{na.action}{A function which indicates what should happen when the data contain \code{NA}s.  Same default behavior as in \code{stats::lm}.}

\item{control}{List of parameters to control the fitting process.  See \strong{Details}.}

\item{model, x, y, qr}{Logical values indicating which data elements should be returned in the output.  These correspond to the response, the covariance matrices, the \code{model.frame}, and the QR decomposition of the hessian of the negative loglikelihood, respectively.  See \code{stats::lm}.}

\item{offset}{Currently ignored, as are \code{offset} terms in the formula.}
}
\value{
An object of class \code{hlm} with the following elements:
\describe{
  \item{\code{coefficients}}{A list with elements \code{beta} and \code{gamma} containing the coefficient MLEs.}
  \item{\code{loglik}}{The loglikelihood at the MLE.}
  \item{\code{df.residual}}{The residual degrees of freedom.}
  \item{\code{iter}}{The number of iterations used in the fitting algorithm.}
  \item{\code{call}}{The matched call.}
  \item{\code{terms}}{A list with elements \code{X} and \code{Z} giving the \code{terms} object for mean and variance models.}
  \item{\code{y}}{If requested, the response used.}
  \item{\code{x}}{If requested, a list with elements \code{X} and \code{Z} giving the mean and variance model matrices uses.}
  \item{\code{model}}{If requested, the \code{model.frame} used.}
  \item{\code{na.action}}{(Where relevant) information returned by \code{model.frame} on the special handling of \code{NA}s.}
  \item{\code{qr}}{If requested, the QR decomposition of the observed Fisher information matrix of size \code{(p+q) x (p+q)}.}
}
}
\description{
Calculates the MLE of a heteroscedastic linear model (HLM) with possibly right-censored responses.  See \strong{Details}.
}
\details{
The heteroscedastic linear model (HLM) is defined as
\deqn{
y_i \mid \boldsymbol{x}_i, \boldsymbol{z}_i \stackrel{\mathrm{ind}}{\sim} \mathcal N\big(\boldsymbol{x}_i'\boldsymbol{\beta}, \exp(\boldsymbol{z}_i'\boldsymbol{\gamma})\big),
}{
y_i | x_i, z_i ~ind N(x_i'\beta, exp(z_i'\gamma)),
}
where for each subject \eqn{i}, \eqn{y_i} is the response, and \eqn{\boldsymbol{x}_i \in \mathbb{R}^p}{x_i \in R^p} and \eqn{\boldsymbol{z}_i \in \mathbb{R}^q}{z_i \in R^q} are mean and variance covariate vectors, respectively.

The \code{formula} term is specified as e.g.,
\preformatted{
y ~ x1 + x2 | z1 + z2
}
where the vertical bar separates mean and variance components.  If no bar is provided, an intercept variance term of the form \code{y ~ x1 + x2 | 1} is assumed, corresponding to the usual linear model (but with a different parametrization).

Right censoring of observations is supported by specifying the response as a two-column matrix, where the first column is the response and the second column is a censoring status indicator with \code{0}: right-censored and \code{1}: uncensored.

Fitting the \code{hlm} model is done by blockwise coordinate ascent, alternatively maximing the mean parameters by weighted least-squares, and the variance parameters either via Fisher scoring or Iteratively Reweighted Least Squares.  When there is right-censoring, these maximization steps are embedded within an Expectation-Conditional-Maximization algorithm.
}
\note{
\strong{Warning:} At present \code{hlm} cannot handle pure LM or LVLM models.  For datasets without censoring, please use the low-level functions \code{\link{lm_fit}} and \code{\link{lvlm_fit}} instead.
}
\examples{
summary(colond) # colon cancer dataset

# fit model without censoring (i.e., ignoring it)
M <- hlm(log(time) ~ . | rx, data = colond)
M # brief display
summary(M) # full summary
vcov(M) # variance estimate

# fit same model but with censoring
hlm(cbind(log(time), status) ~ . | rx, data = colond)
}
\references{
Wang, Y., You, T., and Lysy, M. "A heteroscedastic accelerated failure time model for survival analysis" (2019): \url{https://arxiv.org/abs/1508.05137}.
}
\seealso{
Current methods for \code{hlm} objects are: \code{print}, \code{nobs}, \code{vcov}, and \code{\link{summary}}.
}
