% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_fit.R
\name{hlm_fit}
\alias{hlm_fit}
\title{Low-level fitting function for the uncensored HLM model.}
\usage{
hlm_fit(y, X, Z, method = c("IRLS", "Fisher"), beta0, gamma0,
  maxit = 100, epsilon = 1e-05)
}
\arguments{
\item{y}{Vector of observations of length \code{n}.}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}

\item{method}{Which method to use for fitting the conditional LVLM model.  See \code{\link{lvlm_fit}}.}

\item{beta0}{Optional initial mean parameter vector of length \code{p}.}

\item{gamma0}{Optional initial variance parameter vector of length \code{q}.}

\item{maxit}{Maximum number of iteration of the fitting algorithm (see \strong{Details}).}

\item{epsilon}{Tolerance threshold for termination of the algorithm (see \strong{Details}).}
}
\value{
A list with elements:
\describe{
  \item{\code{beta}}{The MLE of the mean parameters as a vector of length \code{p}.}
  \item{\code{gamma}}{The MLE of the variance parameters as a vector of length \code{q}.}
  \item{\code{loglik}}{The loglikelihood at the final step of the algorithm.}
  \item{\code{iter}}{The number of iterations of the fitting algorithm.}
  \item{\code{tolerance}}{The loglikelihood relative error at the last step.}
}
}
\description{
Calculates the MLE of the coefficients of the heteroscedastic linear model (HLM).  See \strong{Details}.
}
\details{
The heteroscedastic linear model (HLM) is defined as
\deqn{
y_i \mid \boldsymbol{x}_i, \boldsymbol{z}_i \stackrel{\mathrm{ind}}{\sim} \mathcal N\big(\boldsymbol{x}_i'\boldsymbol{\beta}, \exp(\boldsymbol{z}_i'\boldsymbol{\gamma})\big),
}{
y_i | x_i, z_i ~ind N(x_i'\beta, exp(z_i'\gamma)),
}
where for each subject \eqn{i}, \eqn{y_i} is the response, and \eqn{\boldsymbol{x}_i \in \mathbb{R}^p}{x_i \in R^p} and \eqn{\boldsymbol{z}_i \in \mathbb{R}^q}{z_i \in R^q} are mean and variance covariate vectors, respectively.

The low-level function \code{hlm_fit} assumes that the response vector is fully observed (uncensored).  See \code{\link{chlm_fit}} for the corresponding function with censoring, or the higher-level interface \code{\link{hlm}}.

The tuning parameters of the LVLM fitting methods are tuned to their default values in \code{\link{lvlm_fit}}.

\strong{Warning:} This \R wrapper function provides a direct interface to the C++ source code.  Incorrect argument specification may lead to abrupt termination of the \R session.
}
