% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_fit.R
\name{lm_fit}
\alias{lm_fit}
\title{Low-level fitting function for the LM model.}
\usage{
lm_fit(y, X, w)
}
\arguments{
\item{y}{Vector of observations of length \code{n}.}

\item{X}{Mean covariate matrix of size \code{n x p}.}

\item{w}{Optional positive weights vector of length \code{n}.}
}
\value{
MLE estimate of \code{beta} as a vector of length \code{p}.
}
\description{
Calculates the MLE of the coefficients of the usual linear regression model (LM).
}
\details{
The LM model is defined as
\deqn{
\boldsymbol{y}_i \mid \boldsymbol{x}_i \stackrel{\mathrm{ind}}{\sim} \mathcal{N}(\boldsymbol{x}_i'\boldsymbol{\beta}, 1/w_i),
}{
y_i | x_i ~ind N(x_i'\beta, 1/w_i),
}
where for each subject \eqn{i}, \eqn{y_i} is the response, \eqn{\boldsymbol{x}_i \in \mathbb{R}^p}{x_i \in R^p} is the mean covariate vector, and \eqn{w_i} is an optional positive weight (defaults to 1).

\strong{Warning:} This \R wrapper function provides a direct interface to the C++ source code.  Incorrect argument specification may lead to abrupt termination of the \R session.
}
