% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_methods.R
\name{summary.hlm}
\alias{summary.hlm}
\alias{print.summary.hlm}
\title{Summary method for \code{hlm} objects.}
\usage{
\method{summary}{hlm}(object, correlation = FALSE,
  symbolic.cor = FALSE, ...)

\method{print}{summary.hlm}(x, digits = max(3L, getOption("digits") -
  3L), symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{object}{An object of class \code{hlm}.}

\item{correlation}{If \code{TRUE}, calculate the correlation matrix of the estimated parameters.}

\item{symbolic.cor}{If \code{TRUE}, print correlations in symbolic form (see \code{stats::symnum}.}

\item{...}{Further arguments passed to or from other methods.}

\item{x}{An object of class \code{summary.hlm}.}

\item{digits}{The number of significant digits to use when printing.}

\item{signif.stars}{If \code{TRUE}, significance stars are printed for each coefficient.}
}
\value{
A list with elements:
\describe{
  \item{\code{call}}{The call issuing the HLM \code{object}.}
  \item{\code{terms}}{The terms component from \code{object}.}
  \item{\code{df.residual}}{The component from \code{object}.}
  \item{\code{coefficients}}{List of two matrices of coefficients, standard errors, z-values, and p-values.}
  \item{\code{cov.unscaled}}{Estimated covariance matrix of the estimated coefficients.}
  \item{\code{correlation}}{If \code{correlation = TRUE}, the estimate correlations between the estimated coefficients.}
  \item{\code{symbolic.cor}}{If \code{correlation = TRUE}, the value of \code{symbolic.cor}.}
  \item{\code{aic}}{The AIC statistic.}
  \item{\code{iter}}{The component from \code{object}.}
  \item{\code{na.action}}{The component from \code{object}.}
}
}
\description{
Summary method for \code{hlm} objects.
}
