% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_fit.R
\name{lvlm_fit}
\alias{lvlm_fit}
\title{Low-level fitting function for the LVLM model.}
\usage{
lvlm_fit(y2, Z, method = c("IRLS", "Fisher", "LS"), gamma0, maxit = 25,
  epsilon = 1e-08)
}
\arguments{
\item{y2}{Square of response vector of length \code{n}.}

\item{Z}{Variance covariate matrix of size \code{n x q}.}

\item{method}{Which fitting algorithm to use.  See \strong{Details}.}

\item{gamma0}{Initial variance parameter vector of length \code{q}.  If missing a least-squares estimate is used (see \strong{Details}).}

\item{maxit}{Maximum number of iteration of the fitting algorithm (see \strong{Details}).}

\item{epsilon}{Tolerance threshold for termination of the algorithm (see \strong{Details}).}
}
\value{
The MLE (or least-squares estimate) of \code{gamma} as a vector of length \code{q}.
}
\description{
Low-level fitting function for the LVLM model.
}
\details{
the log-variance linear model (LVLM) is defined as...

Three types of fitting algorithms are provided.  \code{method = Fisher} and \code{IRLS} are Fisher Scoring and Iteratively Reweighted Least-Squares MLE-finding algorithms, respectively.  The former is faster while the latter is more stable.  \code{method = LS} is a least-squares estimator, which is the fastest.  It is a consistent estimator but not as efficient as the MLE.

\strong{Warning:} This \R wrapper function provides a direct interface to the C++ source code.  Incorrect argument specification may lead to abrupt termination of the \R session.
}
